% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askChatGPT.R
\name{askChatGPT}
\alias{askChatGPT}
\title{This function is to ask ChatGPT a question and get the answer from it.}
\usage{
askChatGPT(
  text,
  key = NULL,
  model = "gpt-3.5-turbo",
  textForSystem = NULL,
  moreHeaders = c(),
  returnFull = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{text}{character, the question}

\item{key}{the API key for ChatGPT}

\item{model}{The model to use, "gpt-3.5-turbo"}

\item{textForSystem}{character, the extra information for system}

\item{moreHeaders}{NULL}

\item{returnFull}{logical, whether to return a `response` object. Default is FALSE.}

\item{verbose}{logical, whether to show the respoding details of the server. Default is FALSE.}
}
\description{
This function is to ask ChatGPT a question and get the answer from it.
}
\details{
See this https://platform.openai.com/docs/api-reference for more details
}
\examples{
\dontrun{
debugonce(askChatGPT)
askChatGPT(text = "Compose a poem that explains the concept of recursion in programming.",
textForSystem = "You are a poetic assistant, skilled in explaining complex programming concepts with creative flair.")
)
}
}
