% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askGemini.R
\name{askGemini}
\alias{askGemini}
\title{This function is to ask Gemini a question and get the answer from it.}
\usage{
askGemini(text, key = NULL, returnFull = FALSE, verbose = FALSE)
}
\arguments{
\item{text}{character, the question}

\item{key}{the API key for Gemini.}

\item{returnFull}{logical, whether to return a `response` object. Default is FALSE.}

\item{verbose}{logical, whether to show the respoding details of the server. Default is FALSE.}
}
\description{
This function is to ask Gemini a question and get the answer from it.
}
\details{
askGemini works dependent on the Gemini API, which requires the user to create their own API key.
(1) go to this https://ai.google.dev/ address and click `Get API key in Google AI Studio`;\cr
(2) Then select `Get API key`;\cr
(3) Consent the agreement;\cr
(4) Click Create API key;\cr
(5) Copy this API key and define it into `~/.Renviron` file in bash terminal by
`echo 'GEMINIKEY=YOUR_API_KEY' >> ~/.Renviron`, and remember replace `YOUR_API_KEY`
by your real API key that you just created.\cr
Currently, Gemini API is allowed in 60+ countries, which do not include
mainland China, Germany, etc. If you would like to use this you need a proxy,
which is yet to be implemented.
}
\examples{
\dontrun{
# If your key is configured in ~/.Renviron file then you don't need to provide key option in the function.

# Ask Gemini an English question
askGemini("Who are you")

# Ask Gemini a Chinese question
askGemini("你是谁")

# Ask Gemini a French question
askGemini("Qui es-tu")

# Ask Gemini a Spanish question
askGemini("Quién eres")

key = 'YOUR_API_KEY'  # Replace it by your real Gemini API key
# Ask Gemini English questions
askGemini("Who are you", key)
}
}
