% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.AI_response}
\alias{print.AI_response}
\title{print the text that AI model replies.}
\usage{
\method{print}{AI_response}(x, ...)
}
\arguments{
\item{x}{the text that AI model replies}
}
\description{
print the text that AI model replies.
}
