% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2Image.R
\name{text2Image}
\alias{text2Image}
\title{Use DALL-E2 or DALL-E3 model to convert text to a image (or images).}
\usage{
text2Image(
  prompt,
  model = c("dall-e-3", "dall-e-2"),
  n = 1,
  quality = c("standard", "hd"),
  response_format = c("url", "b64_json"),
  size = c("1024x1024", "1792x1024", "1024x1792", "256x256", "512x512"),
  style = c("vivid", "natural"),
  user = NULL,
  output = NULL,
  returnFull = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{prompt}{character, A text description of the desired image(s).
The maximum length is 1000 characters for `dall-e-2` and 4000 characters for
`dall-e-3` model.}

\item{model}{The model to use for image generation. Defaults to `dall-e-3`.}

\item{n}{The number of images to generate. Must be between 1 and 10.
For `dall-e-3`, only `n=1` is supported. Defaults to 1.}

\item{quality}{The quality of the image that will be generated.
`hd` creates images with finer details and greater consistency across the image.
This param is only supported for `dall-e-3`. Defaults to `standard`.}

\item{response_format}{The format in which the generated images are returned.
Must be one of `url` or `b64_json`. URLs are only valid for 60 minutes after the
image has been generated. Defaults to `url`.}

\item{size}{The size of the generated images. Must be one of `256x256`, `512x512`,
or `1024x1024` for `dall-e-2`. Must be one of `1024x1024`, `1792x1024`, or
`1024x1792` for `dall-e-3` models. Defaults to `1024x1024`.}

\item{style}{The style of the generated images. Must be one of `vivid` or `natural`.
Vivid causes the model to lean towards generating hyper-real and dramatic images.
Natural causes the model to produce more natural, less hyper-real looking images.
This param is only supported for `dall-e-3`. Defaults to `vivid`.}

\item{user}{A unique identifier representing your end-user, which can help
OpenAI to monitor and detect abuse.}

\item{output}{Either NULL (default) or character for output picture file name.}

\item{returnFull}{logical, whether to return a `response` object. Default is FALSE.}

\item{verbose}{logical, whether to show the respoding details of the server. Default is FALSE.}
}
\description{
Use DALL-E2 or DALL-E3 model to convert text to a image (or images).
}
\examples{
\dontrun{
# Of note: running text2Image function is relatively more expensive than
# running other text- or audio- related functions.

prompt = "A cute baby sea otter"

# use the default DALL-E3 model to generate a image of 'vivid' type
text2Image(prompt, output = "otter.png")

# use the default DALL-E2 model to generate a image of 'natural' type
text2Image(prompt, model = "dall-e-2", output = "otter2.png")
}

}
