% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2Speach.R
\name{text2Speach}
\alias{text2Speach}
\title{Convert text to an audio file}
\usage{
text2Speach(
  text,
  key = NULL,
  model = c("tts-1", "tts-1-hd"),
  voice = c("alloy", "echo", "fable", "onyx", "nova", "shimmer"),
  response_format = c("mp3", "opus", "aac", "flac", "wav", "pcm"),
  moreHeaders = c(),
  output = "speech",
  speed = 1,
  verbose = FALSE
)
}
\arguments{
\item{text}{character, the question}

\item{key}{the API key for OpenAI}

\item{model}{The model to use, one of c("tts-1", "tts-1-hd")}

\item{voice}{one of c("alloy", "echo", "fable", "onyx", "nova", "shimmer")}

\item{response_format}{one of c("mp3", "opus", "aac", "flac", "wav", "pcm")}

\item{moreHeaders}{NULL}

\item{output}{output file for audio}

\item{speed}{ranging from 0.25 to 4.0. Default is 1.0}

\item{verbose}{logical, whether to show the respoding details of the server. Default is FALSE.}
}
\description{
Convert text to an audio file
}
\examples{
\dontrun{

text2Speach("Compose a poem that explains the concept of recursion in programming.")
text2Speach("Convert this into an English monologue.",
             output = "English.mp3")
text2Speach("把这个转换成中文普通话.", voice = "nova",
             output = "Chinese.mp3")
text2Speach("Convertissez ceci en un monologue anglais", output = "French.mp3")
text2Speach("把这个English conversation转换成中文对话.", voice = "nova",
             output = "Mix.mp3")
}

}
